pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.61.0",
    "pandas>=3.0.0",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.3",
    "yfinance>=1.1.0",
]
